verdigris: clippy cleanup (auto-fix pass + manual style/correctness fixes)
- cargo clippy --fix --workspace: mechanical simplifications (needless_lifetimes,
  useless_conversion, redundant_closure, manual_is_multiple_of,
  unnecessary_cast, needless_question_mark, unnecessary_lazy_evaluations,
  int_plus_one, collapsible_if, several boolean simplifications).
- Three `!(x > 0.0)`/`!(x >= 0.0)` checks (gas/cell.rs delta share, ffi
  reactor.rs's react_step/rate_relax tick and rate validation) are
  deliberately negated comparisons so a NaN input is also rejected, which
  `x <= 0.0` would not catch; annotated with `#[allow(neg_cmp_op_on_partial_ord)]`
  and a comment instead of "fixing" them into a behavior change.
- gas/constants.rs: four `/// SECTION` doc comments that don't document the
  following item (clippy::empty_line_after_doc_comments) -> plain `//`.
- structural.rs: three seed-mixing hex constants regrouped to clippy's
  preferred nibble grouping (unusual_byte_groupings; same values).
- content.rs/contract.rs: two genuinely-identical if/else arms merged
  (clippy::if_same_then_else) -- confirmed both produced the same value/expr.

`cargo test --workspace --target i686-pc-windows-msvc` (including doctests)
is green. Remaining clippy output is dominated by ~50 pre-existing
`dead_code` warnings in domains/layout/src/station_layout/{structural,content}.rs
(functions confirmed unreachable by rustc, left in place rather than bulk-
deleted this pass -- see doc/rewrite/rust_core.md's audit notes) plus a
handful of `too_many_arguments`/`type_complexity` warnings on internal
geometry-heavy generator functions, not fixed here to avoid touching many
call sites' signatures under a "behavior must stay identical" constraint.

+++ verdigris/domains/layout/src/station_layout/render.rs
@@ -577,8 +577,8 @@ fn blit_tile_offset(
         i32::try_from(height).unwrap_or(i32::MAX) - (i32::from(point.y) + 1) * 32 - offset_y;
     for y in 0..32 {
         for x in 0..32 {
-            let target_x = origin_x + x as i32;
-            let target_y = origin_y + y as i32;
+            let target_x = origin_x + x;
+            let target_y = origin_y + y;
             if target_x < 0 || target_y < 0 || target_x >= width as i32 || target_y >= height as i32
             {
                 continue;
